modified:   expgram/NGram.cpp 	modified:   expgram/NGramCounts.cpp 		use of new lockfree queue
	new file:   progs/ngram_counts_estimate_mpi.cpp
	modified:   progs/ngram_counts_impl.hpp
	modified:   progs/ngram_counts_index_mpi.cpp
		use of new lockfree queue

	modified:   progs/tokenize_en.cpp
		added Google 5-gram workaround

	modified:   succinct_db/sample_succinct_trie_db.cpp
	modified:   succinct_db/succinct_hash_db.hpp
	modified:   succinct_db/succinct_trie.hpp
	modified:   succinct_db/succinct_trie_db.hpp
		check for potential error

	modified:   utils/bithack.hpp
		bugfix for bit-count

	modified:   utils/lockfree_list_queue.hpp
		updated list-version of lockfree queue

	modified:   utils/map_file.hpp
		updated for new open

	new file:   utils/mpi_traits.hpp
		added mpi-traits to check for datatype

	modified:   utils/vertical_coded_device.hpp
		bugfix and removed block-file

+++ utils/bithack.hpp
@@ -315,7 +315,7 @@ namespace utils
       static const uint64_t X5 = uint64_t(X4) + (uint64_t(X4) >> 16);
       static const uint64_t X6 = uint64_t(X5) + (uint64_t(X5) >> 32) & 0x0000007F;
     public:
-      static const size_t result = size_t(X6);
+      static const uint64_t result = size_t(X6);
     };
     
     // floor_log2
